refactor frontend better logical separation between composable and component
